Moved LFGRecyclerAdapter into it's own class (swipetorefresh currently broken)
